refactor(search applications): rename page template file
