+++ x-pack/plugins/enterprise_search/public/applications/applications/components/engine/engine_view.tsx
@@ -24,7 +24,7 @@ import {
   SearchApplicationContentTabs,
 } from '../../routes';
 
-import { EnterpriseSearchEnginesPageTemplate } from '../layout/engines_page_template';
+import { EnterpriseSearchEnginesPageTemplate } from '../layout/page_template';
 import { DeleteSearchApplicationModal } from '../search_applications/delete_search_application_modal';
 
 import { EngineConnect } from './engine_connect/engine_connect';
