+++ x-pack/plugins/enterprise_search/public/applications/applications/components/layout/page_template.tsx
