+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/search_applications_list.tsx
@@ -37,7 +37,7 @@ import {
 } from '../../../shared/licensing_callout/licensing_callout';
 
 import { SEARCH_APPLICATIONS_PATH, SEARCH_APPLICATION_CREATION_PATH } from '../../routes';
-import { EnterpriseSearchEnginesPageTemplate } from '../layout/engines_page_template';
+import { EnterpriseSearchEnginesPageTemplate } from '../layout/page_template';
 
 import { EmptySearchApplicationsPrompt } from './components/empty_search_applications_prompt';
 import { SearchApplicationsListTable } from './components/tables/search_applications_table';
